small fixes around the tree

+++ src/web/Utils.h
@@ -55,7 +55,7 @@ inline bool startsWith(const char *a, const char *b, int n) {
 
 inline int length(const std::stringstream& s) {
 #ifndef WT_CNOR
-  return s.rdbuf()->in_avail();
+  return (int)s.rdbuf()->in_avail();
 #else
   return 0;
 #endif
